requirements.txt
langchain
numpy
matplotlib
pandas
python-dotenv
langchain-openai
langchain-core
langchain-community
bs4
faiss-cpu
pypdf
Arxiv
fitz
pymupdf
wikipedia
lxml
langchain_huggingface
langchain_groq
langgraph 
typing_extensions   
langgraph-cli[inmem]
phidata
yfinance
packaging
duckduckgo-search
fastapi
uvicorn
groq
sqlalchemy
pgvector
psycopg[binary]
pypdf
streamlit
google-generativeai
duckduckgo-search
PyPDF2
google-generativeai
langchain_google_genai
transformers
torch
requests
langchain-community
mysql-connector-python 
langgraph
langchain_groq
tiktoken
langchain.tools
pymupdf